Refactoring styles
