+++ vecep/src/views/contents/exercises/style.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const ExercisesContainer = styled.div`
+display: flex;
+flex-direction: column;
+align-items: center;
+margin-bottom: 50px;
+`;
+
+export default ExercisesContainer;
