--- vecep/src/views/contents/style.js
@@ -1,12 +1,0 @@
-import styled from 'styled-components'
-
-const Batatinha = styled.div`
-`
-
-const Image = styled.image`
-    & > img {
-        height: 230px;
-    }
-`
-
-export { Batatinha, Image }
